Identity -> SimpleIdentity
Identity class must exist as an alias in order to read it from session
